+++ qunit/tests_phoneru.js
@@ -26,8 +26,13 @@ define([
 
 		testmask.focus();
 
-		$.each(Inputmask.prototype.defaults.aliases.phoneru.phoneCodes, function (ndx, lmnt) {
-			var expected = lmnt.mask.replace(/#/g, "0"), input = expected;
+		$.each(Inputmask.prototype.defaults.aliases.phoneru.phoneCodes.slice(0,10), function (ndx, lmnt) {
+			var ndx = 1, input, expected = lmnt.mask;
+			while (expected.match(/#/)) {
+				expected = expected.replace(/#/, ndx++);
+				if (ndx > 9) ndx = 1;
+			}
+			input = expected;
 			//input = input.replace(/\+/g, "");
 			input = input.replace(/\(/g, "");
 			input = input.replace(/\)/g, "");